Move the `isNodeJS`-helper into the `src/shared/util.js` file
With the changes in the previous patch the `isNodeJS`-helper no longer needs to live in its own file, which helps get rid of a closure in the *built* files.
