+++ test/unit/clitests_helper.js
@@ -13,8 +13,11 @@
  * limitations under the License.
  */
 
-import { setVerbosityLevel, VerbosityLevel } from "../../src/shared/util.js";
-import { isNodeJS } from "../../src/shared/is_node.js";
+import {
+  isNodeJS,
+  setVerbosityLevel,
+  VerbosityLevel,
+} from "../../src/shared/util.js";
 
 // Sets longer timeout, similar to `jasmine-boot.js`.
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
