+++ test/unit/display_svg_spec.js
@@ -16,7 +16,7 @@
 
 import { buildGetDocumentParams } from "./test_utils.js";
 import { getDocument } from "../../src/display/api.js";
-import { isNodeJS } from "../../src/shared/is_node.js";
+import { isNodeJS } from "../../src/shared/util.js";
 import { SVGGraphics } from "../../src/display/svg.js";
 
 const XLINK_NS = "http://www.w3.org/1999/xlink";
