+++ test/unit/display_utils_spec.js
@@ -13,6 +13,7 @@
  * limitations under the License.
  */
 
+import { bytesToString, isNodeJS } from "../../src/shared/util.js";
 import {
   DOMCanvasFactory,
   DOMSVGFactory,
@@ -21,8 +22,6 @@ import {
   isValidFetchUrl,
   PDFDateString,
 } from "../../src/display/display_utils.js";
-import { bytesToString } from "../../src/shared/util.js";
-import { isNodeJS } from "../../src/shared/is_node.js";
 
 describe("display_utils", function () {
   describe("DOMCanvasFactory", function () {
