+++ test/unit/xfa_tohtml_spec.js
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { isNodeJS } from "../../src/shared/is_node.js";
+import { isNodeJS } from "../../src/shared/util.js";
 import { XFAFactory } from "../../src/core/xfa/factory.js";
 
 describe("XFAFactory", function () {
